pyproject.toml
[project]
name = "judgely"
version = "0.1.0"
description = "LLM Court - Judgment Day"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "datasets>=3.6.0",
    "ollama>=0.5.1",
    "openai>=1.86.0",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ty>=0.0.1a11",
]
